services/gitlab-api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-gitlab>=3.0.0
httpx>=0.25.0